requirements.txt

flask
flask-cors
llama-cpp-python
